# Improve performance of ByteArraysEqual