@davidfowl 2019-04-27
We should delete this method and use CryptographicOperations.FixedTimeEquals instead.

cc @blowdart @GrabYourPitchforks