@blowdart 2019-04-27
Agreed, @davidfowl please open another issue for that.

The function is deliberately not optimised. Closing.