@meziantou 2019-04-27
Thanks for the feedback! I've created a new pull request #9798 with FixedTimeEquals.

BTW, FixedTimeEquals uses a substract operator which is a little more performant than the xor operator (at least on my computer).